+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Deleg.hs
@@ -34,11 +34,10 @@ import Cardano.Ledger.Conway.TxCert (
   ConwayDelegCert (ConwayDelegCert, ConwayRegCert, ConwayRegDelegCert, ConwayUnRegCert),
   Delegatee (DelegStake, DelegStakeVote, DelegVote),
  )
-import Cardano.Ledger.Core (Era (EraCrypto), EraPParams, EraRule)
+import Cardano.Ledger.Core (Era (EraCrypto), EraPParams, EraRule, PParams)
 import Cardano.Ledger.Credential (Credential)
 import Cardano.Ledger.Keys (KeyRole (Staking))
 import Cardano.Ledger.Shelley.LedgerState (DState (..))
-import Cardano.Ledger.Shelley.Rules (DelegEnv (DelegEnv))
 import qualified Cardano.Ledger.UMap as UM
 import Control.DeepSeq (NFData)
 import Control.Monad (forM_)
@@ -64,9 +63,9 @@ import NoThunks.Class (NoThunks)
 
 data ConwayDelegPredFailure era
   = IncorrectDepositDELEG !Coin
-  | StakeKeyAlreadyRegisteredDELEG !(Credential 'Staking (EraCrypto era))
+  | StakeKeyRegisteredDELEG !(Credential 'Staking (EraCrypto era))
   | StakeKeyNotRegisteredDELEG !(Credential 'Staking (EraCrypto era))
-  | StakeKeyHasNonZeroAccountBalanceDELEG !Coin
+  | StakeKeyHasNonZeroRewardAccountBalanceDELEG !Coin
   | DRepAlreadyRegisteredForStakeKeyDELEG !(Credential 'Staking (EraCrypto era))
   | WrongCertificateTypeDELEG
   deriving (Show, Eq, Generic)
@@ -80,12 +79,12 @@ instance Era era => EncCBOR (ConwayDelegPredFailure era) where
     encode . \case
       IncorrectDepositDELEG mCoin ->
         Sum (IncorrectDepositDELEG @era) 1 !> To mCoin
-      StakeKeyAlreadyRegisteredDELEG stakeCred ->
-        Sum (StakeKeyAlreadyRegisteredDELEG @era) 2 !> To stakeCred
+      StakeKeyRegisteredDELEG stakeCred ->
+        Sum (StakeKeyRegisteredDELEG @era) 2 !> To stakeCred
       StakeKeyNotRegisteredDELEG stakeCred ->
         Sum (StakeKeyNotRegisteredDELEG @era) 3 !> To stakeCred
-      StakeKeyHasNonZeroAccountBalanceDELEG mCoin ->
-        Sum (StakeKeyHasNonZeroAccountBalanceDELEG @era) 4 !> To mCoin
+      StakeKeyHasNonZeroRewardAccountBalanceDELEG mCoin ->
+        Sum (StakeKeyHasNonZeroRewardAccountBalanceDELEG @era) 4 !> To mCoin
       DRepAlreadyRegisteredForStakeKeyDELEG stakeCred ->
         Sum (DRepAlreadyRegisteredForStakeKeyDELEG @era) 5 !> To stakeCred
       WrongCertificateTypeDELEG ->
@@ -94,9 +93,9 @@ instance Era era => EncCBOR (ConwayDelegPredFailure era) where
 instance Era era => DecCBOR (ConwayDelegPredFailure era) where
   decCBOR = decode $ Summands "ConwayDelegPredFailure" $ \case
     1 -> SumD IncorrectDepositDELEG <! From
-    2 -> SumD StakeKeyAlreadyRegisteredDELEG <! From
+    2 -> SumD StakeKeyRegisteredDELEG <! From
     3 -> SumD StakeKeyNotRegisteredDELEG <! From
-    4 -> SumD StakeKeyHasNonZeroAccountBalanceDELEG <! From
+    4 -> SumD StakeKeyHasNonZeroRewardAccountBalanceDELEG <! From
     5 -> SumD DRepAlreadyRegisteredForStakeKeyDELEG <! From
     6 -> SumD WrongCertificateTypeDELEG
     n -> Invalid n
@@ -107,14 +106,14 @@ instance
   ( EraPParams era
   , State (EraRule "DELEG" era) ~ DState era
   , Signal (EraRule "DELEG" era) ~ ConwayDelegCert (EraCrypto era)
-  , Environment (EraRule "DELEG" era) ~ DelegEnv era
+  , Environment (EraRule "DELEG" era) ~ PParams era
   , EraRule "DELEG" era ~ ConwayDELEG era
   ) =>
   STS (ConwayDELEG era)
   where
   type State (ConwayDELEG era) = DState era
   type Signal (ConwayDELEG era) = ConwayDelegCert (EraCrypto era)
-  type Environment (ConwayDELEG era) = DelegEnv era
+  type Environment (ConwayDELEG era) = PParams era
   type BaseM (ConwayDELEG era) = ShelleyBase
   type PredicateFailure (ConwayDELEG era) = ConwayDelegPredFailure era
   type Event (ConwayDELEG era) = ConwayDelegEvent era
@@ -124,7 +123,7 @@ instance
 conwayDelegTransition :: forall era. EraPParams era => TransitionRule (ConwayDELEG era)
 conwayDelegTransition = do
   TRC
-    ( DelegEnv _slot _ptr _acnt pp
+    ( pp
       , dState@DState {dsUnified}
       , c
       ) <-
@@ -133,34 +132,29 @@ conwayDelegTransition = do
   case c of
     ConwayRegCert stakeCred sMayDeposit -> do
       forM_ sMayDeposit $ checkDepositAgainstPParams ppKeyDeposit
-      checkStakeKeyNotAlreadyRegistered stakeCred dsUnified
-      pure $ dState {dsUnified = acceptDepositForStakeKey stakeCred dsUnified ppKeyDeposit}
+      dsUnified' <- checkAndAcceptDepositForStakeCred stakeCred ppKeyDeposit dsUnified
+      pure $ dState {dsUnified = dsUnified'}
     ConwayUnRegCert stakeCred sMayDeposit -> do
-      checkStakeKeyIsAlreadyRegistered stakeCred dsUnified
-      checkStakeKeyHasZeroBalance stakeCred dsUnified
+      checkStakeKeyIsRegistered stakeCred dsUnified
+      checkStakeKeyHasZeroRewardBalance stakeCred dsUnified
       forM_ sMayDeposit $ checkDepositAgainstPaidDeposit stakeCred dsUnified
-      pure $
-        dState
-          { dsUnified = UM.domDeleteAll (Set.singleton stakeCred) dsUnified
-          }
+      pure $ dState {dsUnified = UM.domDeleteAll (Set.singleton stakeCred) dsUnified}
     ConwayDelegCert stakeCred delegatee -> do
-      checkStakeKeyIsAlreadyRegistered stakeCred dsUnified
-      pure $
-        dState
-          { dsUnified = processDelegation stakeCred delegatee dsUnified
-          }
+      checkStakeKeyIsRegistered stakeCred dsUnified
+      pure $ dState {dsUnified = processDelegation stakeCred delegatee dsUnified}
     ConwayRegDelegCert stakeCred delegatee deposit -> do
-      deposit == ppKeyDeposit ?! IncorrectDepositDELEG deposit
-      checkStakeKeyNotAlreadyRegistered stakeCred dsUnified
-      pure $
-        dState
-          { dsUnified =
-              processDelegation stakeCred delegatee $
-                acceptDepositForStakeKey stakeCred dsUnified ppKeyDeposit
-          }
+      checkDepositAgainstPParams ppKeyDeposit deposit
+      dsUnified' <- checkAndAcceptDepositForStakeCred stakeCred deposit dsUnified
+      pure $ dState {dsUnified = processDelegation stakeCred delegatee dsUnified'}
   where
-    acceptDepositForStakeKey stakeCred dsUnified ppKeyDeposit =
-      UM.RewDepUView dsUnified UM.∪ (stakeCred, UM.RDPair (UM.CompactCoin 0) (UM.compactCoinOrError ppKeyDeposit))
+    -- Whenever we want to accept new deposit, we must always check if the stake credential isn't already registered.
+    checkAndAcceptDepositForStakeCred stakeCred deposit dsUnified = do
+      checkStakeKeyNotRegistered stakeCred dsUnified
+      -- This looks like it should have been a right-biased union, so that the (reward, deposit) pair would be inserted
+      -- (or overwritten) in the UMap. But since we are sure that the stake credential isn't a member yet
+      -- it will still work. The reason we cannot use a right-biased union here is because UMap treats deposits specially
+      -- in right-biased unions, and is unable to accept new deposits.
+      pure $ UM.RewDepUView dsUnified UM.∪ (stakeCred, UM.RDPair (UM.CompactCoin 0) (UM.compactCoinOrError deposit))
     delegStake stakeCred sPool dsUnified =
       UM.SPoolUView dsUnified UM.⨃ Map.singleton stakeCred sPool
     delegVote stakeCred dRep dsUnified =
@@ -174,10 +168,10 @@ conwayDelegTransition = do
       deposit == ppKeyDeposit ?! IncorrectDepositDELEG deposit
     checkDepositAgainstPaidDeposit stakeCred dsUnified deposit =
       Just deposit == fmap (UM.fromCompact . UM.rdDeposit) (UM.lookup stakeCred $ UM.RewDepUView dsUnified) ?! IncorrectDepositDELEG deposit
-    checkStakeKeyNotAlreadyRegistered stakeCred dsUnified =
-      UM.notMember stakeCred (UM.RewDepUView dsUnified) ?! StakeKeyAlreadyRegisteredDELEG stakeCred
-    checkStakeKeyIsAlreadyRegistered stakeCred dsUnified =
+    checkStakeKeyNotRegistered stakeCred dsUnified =
+      UM.notMember stakeCred (UM.RewDepUView dsUnified) ?! StakeKeyRegisteredDELEG stakeCred
+    checkStakeKeyIsRegistered stakeCred dsUnified =
       UM.member stakeCred (UM.RewDepUView dsUnified) ?! StakeKeyNotRegisteredDELEG stakeCred
-    checkStakeKeyHasZeroBalance stakeCred dsUnified =
+    checkStakeKeyHasZeroRewardBalance stakeCred dsUnified =
       let mReward = UM.rdReward <$> UM.lookup stakeCred (UM.RewDepUView dsUnified)
-       in forM_ mReward $ \r -> r == mempty ?! StakeKeyHasNonZeroAccountBalanceDELEG (UM.fromCompact r)
+       in forM_ mReward $ \r -> r == mempty ?! StakeKeyHasNonZeroRewardAccountBalanceDELEG (UM.fromCompact r)